gps: Add loc_eng_set_privacy function to toggle GPS privacy lock
Privacy lock is enabled on loc_eng_init, disabled on loc_eng_cleanup,
mimicking the behavior of TW.  Requires system property
ro.gps.set_privacy=1 to enable/disable privacy lock.

The privacy lock state is non-volatile, and control of the lock is needed
to ensure that the GPS works and is responsive to requests. Otherwise the
GPS may silently cease to work depending on its last state in TW.

Change-Id: I7401cc9bec75430646944971d20270f97a5043bc